requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
umap-learn>=0.5.3
bertopic>=0.15.0
wordcloud>=1.9.0
anthropic>=0